Fixed regression text/plain is not copied
