+++ e2e/custom.spec.ts
@@ -1,4 +1,4 @@
-import { test, expect, Page } from "@playwright/test";
+import { test, expect } from "@playwright/test";
 import {
   createSelection,
   getSelection,
@@ -301,129 +301,5 @@ test.describe("smoke node", () => {
     expect(await getSelection(editable)).toEqual(
       createSelection({ offset: nodeOffset })
     );
-  });
-});
-
-test.describe("html paste", () => {
-  test.beforeEach(async ({ context, browserName }) => {
-    // https://github.com/microsoft/playwright/issues/13037#issuecomment-1078208810
-    test.skip(browserName !== "chromium");
-    await context.grantPermissions(["clipboard-read", "clipboard-write"]);
-  });
-
-  const writeHTML = async (page: Page, value: string) => {
-    await page.evaluate(
-      (t) =>
-        navigator.clipboard.write([
-          new ClipboardItem({
-            "text/html": new Blob([t], { type: "text/html" }),
-          }),
-        ]),
-      value
-    );
-  };
-
-  test("single paragraph root", async ({ page }) => {
-    await page.goto(storyUrl("basics-custom--image"));
-
-    const editable = await getEditable(page);
-
-    await editable.focus();
-
-    const plain = `
-export const editable = (
-  element: HTMLElement,
-  { readonly, nodes, onChange }: EditableOptions`;
-    const html = `<meta charset='utf-8'><div><br><div><span>export</span><span> </span><span>const</span><span> </span><span>editable</span><span> </span><span>=</span><span> (</span></div><div><span>  </span><span>element</span><span>:</span><span> </span><span>HTMLElement</span><span>,</span></div><div><span>  { </span><span>readonly</span><span>, </span><span>nodes</span><span>, </span><span>onChange</span><span> }</span><span>:</span><span> </span><span>EditableOptions</span></div></div>`;
-
-    await writeHTML(page, html);
-
-    // Delete all
-    await page.keyboard.press(`ControlOrMeta+A`);
-    await page.keyboard.press(`Backspace`);
-
-    await page.keyboard.press(`ControlOrMeta+V`);
-    expect((await getText(editable)).join("\n")).toEqual(plain);
-  });
-
-  test("multi paragraph root", async ({ page }) => {
-    await page.goto(storyUrl("basics-custom--image"));
-
-    const editable = await getEditable(page);
-
-    await editable.focus();
-
-    const plain = `#17
-#6`;
-    const html = `<meta charset='utf-8'><p>#17</p><p>#6</p>`;
-
-    await writeHTML(page, html);
-
-    // Delete all
-    await page.keyboard.press(`ControlOrMeta+A`);
-    await page.keyboard.press(`Backspace`);
-
-    await page.keyboard.press(`ControlOrMeta+V`);
-    expect((await getText(editable)).join("\n")).toEqual(plain);
-  });
-
-  test("single inline root", async ({ page }) => {
-    await page.goto(storyUrl("basics-custom--image"));
-
-    const editable = await getEditable(page);
-
-    await editable.focus();
-
-    const plain = `#17
-#6`;
-    const html = `<meta charset='utf-8'><span>#17<br ><em>#6</em></span>`;
-    await writeHTML(page, html);
-
-    // Delete all
-    await page.keyboard.press(`ControlOrMeta+A`);
-    await page.keyboard.press(`Backspace`);
-
-    await page.keyboard.press(`ControlOrMeta+V`);
-    expect((await getText(editable)).join("\n")).toEqual(plain);
-  });
-
-  test("multi inline root", async ({ page }) => {
-    await page.goto(storyUrl("basics-custom--image"));
-
-    const editable = await getEditable(page);
-
-    await editable.focus();
-
-    const plain = `#17
-#6`;
-    const html = `<meta charset='utf-8'><a>#17</a><br ><a>#6</a>`;
-    await writeHTML(page, html);
-
-    // Delete all
-    await page.keyboard.press(`ControlOrMeta+A`);
-    await page.keyboard.press(`Backspace`);
-
-    await page.keyboard.press(`ControlOrMeta+V`);
-    expect((await getText(editable)).join("\n")).toEqual(plain);
-  });
-
-  test("table root", async ({ page }) => {
-    await page.goto(storyUrl("basics-custom--image"));
-
-    const editable = await getEditable(page);
-
-    await editable.focus();
-
-    const plain = `    const html = clipboardData.getData("text/html");
-    if (html) {`;
-    const html = `<meta charset='utf-8'><table><tbody><tr><td><span>    <span>const</span> <span>html</span> <span>=</span> <span>clipboardData</span><span>.</span><span>getData</span><span>(</span><span>"text/html"</span><span>)</span><span>;</span></span></td></tr><tr><td></td><td></td><td><button><svg aria-hidden="true" height="16" viewBox="0 0 16 16" version="1.1" width="16"><path></path></svg></button><span>    <span>if</span> <span>(</span><span>html</span><span>)</span> <span>{</span></span></td></tr></tbody></table>`;
-    await writeHTML(page, html);
-
-    // Delete all
-    await page.keyboard.press(`ControlOrMeta+A`);
-    await page.keyboard.press(`Backspace`);
-
-    await page.keyboard.press(`ControlOrMeta+V`);
-    expect((await getText(editable)).join("\n")).toEqual(plain);
   });
 });
